pkg/proc: Move SetCurrentBreakpoint up (pt 1)
This patch moves some of the SetCurrentBreakpoint calls that were
happening in proc/native and proc/gdbserial up to proc and removes some
duplication.

When doing this I ran into an issue with Windows. For some reason I
don't fully understand, unlike the other backends, if you blindly
iterate over each thread and call SetCurrentBreakpoint it breaks Windows
in unexpected ways which manifests itself in the `StepConcurrent` tests.
This is resolved by returning not only the initial thread that we first
received the trap from, but all other threads that may have hit a
breakpoint as well. And then we only call SetCurrentBreakpoints on those
threads.