feature(examinemem): drop format `address`
When format the data at some address, first determine the byte order
then calculate the number. Finally format it and display it. Here, I
suggest use 1byte zero padding to format the number, it will be shorter
and looks a little better, for example, a 8-byte hexidecimal, 0x12345678
may be better than 0x0000000012345678.