service/dap: complete step requests by skipping bps in other goroutines
Other goroutines may encounter breakpoints during a step request on
a certain goroutine. When this happens, state.NextInProgress is set
to true. If another step request is issued, then we get a "next while
nexting" error from the debugger. Instead we have two options:

1. skip all breakpoints and send output events to notify the user.
2. cancel next and notify user when the breakpoint is hit.

The terminal package implements (1), and this makes sense for dlv dap
as well because when users hit next, they are choosing to focus on a
particular goroutine. In order to get the behavior where they stop when
a breakpoint is hit, they can simply add a breakpoint to the next line,
and then run continue.