service/dap: cancel next before sending stopped event
When 'NextInProgress' is set, sending a 'next', 'step', or 'stepOut'
request will result in a 'next while nexting' error. This change
cancels the 'NextInProgress' before sending the stopped event, so
users can continue execution as desired.

The other option for handling 'NextInProgress' would be to instead
skip any breakpoints that are encountered and notifying the user.
I chose to do cancel next instead because it is the simpler to
implement of the two choices and is easy to understand.