dap: auto-load partially loaded nested variables
Keep track of whether a variable is partially loaded in the variable
handle. When a partial variable is requested (since user manually
expands the variable through UI or debug console), try to load with
the expression, and replace the partially loaded previous variable
with the newly updated one.

Unfortunately, the newly loaded variable may be still partial
if the struct has many fields, or the variable is a map, array, or
struct. They will be addressed separately - probably by utilizing
pagination UI.

Keep track of the goid and frame information, so they are
available when the partially loaded variable is auto-loaded.