pkg/proc: make sure that the pending signal is processed
This attempts to solve the issue that a manual halt request
is sent while stopped by checking if there is a pending
signal, and resuming execution to receive the signal.

This also includes tests in pkg/proc and service/dap
for the issue this is attempting to solve.

Updates go-delve/delve#2685