service/dap: fix closing closed channel bug
Our previous implementation of waiting on resumeRequestLoop was not
safe, because the channel may be closed several times. This change
removes this bug by instead having two channels to track the two
conditions that we are managing, whether the command has started running
and whether the command has completed (on error or other).

Fixes go-delve/delve#2757